changed it into a component
