+++ react-app/src/components/Comments/comment.js
@@ -1,0 +1,133 @@
+import React, {useState,useRef,useEffect} from 'react'
+import { useSelector, useDispatch } from "react-redux";
+import {useDetectOutsideClick} from "../../services/detectOutsideClick"
+import { updateComments, deleteComment } from "../../Store/comments";
+import blankHeart from "../../images/icons/insta_heart_blank_icon.png";
+import redHeart from "../../images/icons/insta_heart_red_icon.png";
+
+
+export default function CommentContent({ comment }) {
+    const dispatch = useDispatch();
+    const [isLiked, setIsLiked] = useState(false);
+    const user = useSelector((state) => state.session.user);
+    const dropdownRef = useRef(null);
+    const [isActive, setIsActive] = useDetectOutsideClick(dropdownRef, false);
+
+
+
+
+    const likeHandler = () => {
+        // const like = { userId: user.id, postId: post.id };
+        setIsLiked(!isLiked);
+        // dispatch(postLike(like));
+      };
+      let activeElement
+
+
+    const handleEnter = async(e,commentId) => {
+        // console.log(commentId)
+        activeElement = document.querySelector(`#comment-${comment.id}`)
+
+        let content = activeElement.firstChild.nodeValue
+        console.log(content)
+        if (e.key === "Enter") {
+          e.preventDefault()
+            await dispatch(updateComments(commentId, content));
+            activeElement.setAttribute("contenteditable", false);
+            activeElement.classList.remove('highlight')
+        }
+      }
+      
+      const showDropMenu = (comment, user) => {
+          if (comment.userId !== user.id) return
+        setIsActive(!isActive);
+        activeElement = document.querySelector(`#comment-${comment.id}`)
+    
+        // activeElement.classList.add('active')
+        console.log(activeElement)
+    
+      }
+      const removeComment = async (commentId) => {
+		await dispatch(deleteComment(commentId));
+	};
+
+
+    const editActionHandler = (e, comment) => {
+        activeElement = document.querySelector(`#comment-${comment.id}`)
+
+        activeElement.setAttribute("contenteditable", true);
+        activeElement.classList.add('highlight');
+        activeElement.focus()
+    }
+
+    return (
+        <>
+            <div className="comments__container menu-container">
+                <div className="comments__user-comment">
+                    <div className="comment__username">{comment.username}</div>
+                <div
+                    id={`comment-${comment.id}`}
+                    className="comment__content menu-trigger"
+                    suppressContentEditableWarning="true"
+                    contentEditable='false'
+                    onKeyPress={(e)=>handleEnter(e,comment.id)}
+                    onClick={() => showDropMenu(comment,user)}
+                            >
+                            {comment.content}
+                            
+                            {comment.userId === user.id && (
+                                <>
+                                <nav
+                                    ref={dropdownRef}
+                                    className={`menu abs ${isActive ? "active" : "inactive"}`}
+                                    // className={`menu abs`}
+                                >
+                                        <ul>
+                                            <li>
+                                                <div
+                                                    className=""
+                                                    suppressContentEditableWarning="true"
+                                                    contentEditable="false"
+                                                    onClick={() => removeComment(comment.id)}
+                                                >
+                                                    Delete
+                                                </div>
+                                            </li>
+                                            <li>
+                                                <div
+                                                    className=""
+                                                    suppressContentEditableWarning="true"
+                                                    contentEditable="false"
+                                                    onClick={(e) => editActionHandler(e,comment)}
+                                                >
+                                                    Edit
+                                                </div>
+                                            </li>
+                                        </ul>
+                                    </nav>
+                                </>
+                            )}
+                        <div className='commentLikeContainer'>
+                            
+               </div>
+                        <img
+                            className='commentLike-icon'
+        src={isLiked ? redHeart : blankHeart}
+        alt='post like button'
+        onClick={() => likeHandler()}
+        />
+
+                    </div>
+                </div>
+            </div>
+
+            
+        </>
+    )
+
+
+
+
+
+
+}
