+++ react-app/src/components/Comments/dropdown.css
@@ -57,19 +57,19 @@
     z-index: 34
   }
   
-  /* .menu.active {
-    opacity: 1;
-    visibility: visible;
-    transform: translateY(0);
-
-  } */
-
-  .comment__content.active nav {
+  .menu.active {
     opacity: 1;
     visibility: visible;
     transform: translateY(0);
 
   }
+
+  /* .comment__content.active nav {
+    opacity: 1;
+    visibility: visible;
+    transform: translateY(0);
+
+  } */
 
 
   .highlight {
